agents/autonomous_agent.py
import json
from memory_stream import MemoryStream
from memory import Observation, Plan, Ideal, Reflection
import uuid

class AutonomousAgent:
    def __init__(self, api_communication, instance_id, agent_roles=None, agent_mission=None):
        self.api_communication = api_communication
        self.instance_id = instance_id
        self.instance_data = self.api_communication.get_instance_data(instance_id)
        self.agent_roles = agent_roles or {}
        self.agent_mission = agent_mission or ''
        self.memory_stream = MemoryStream()

    def get_conversation_history(self):
        return self.api_communication.get_conversation_history(self.instance_id)

    def process_message(self, message_data):
        # Process the incoming message and update the agent's memory tree accordingly.
        command = self.parse_command(message_data)
        self.handle_command(command)

    def communicate(self, message_data):
        # Fetch conversation history from the API
        conversation_history = self.get_conversation_history()
        # Process the message and update the memory tree
        self.process_message(message_data)

# Rest of the class code remains the same...
